fix, build(pinia): rename components & add pinia - changed component file structures - installed pinia - added pinia stores
