+++ components/search/area/MainTitle.vue
